requirements.txt
Django>=5.0
djangorestframework
folium
django-cors-headers